api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
slowapi==0.1.9
pydantic==2.5.0
pydantic-settings==2.1.0
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
tensorflow==1.13.1
scikit-image==0.19.3
scipy==1.10.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
structlog==23.2.0

